pyproject.toml
[project]
name = "rossmann-mlops"
version = "0.1.0"
description = "MLOps pipeline for Rossmann sales forecasting"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.0",
    "pydantic>=2.0.0",
    "joblib>=1.4.2",
    "numpy>=2.1.0",
    "pandas>=2.2.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.5.0",
    "uvicorn>=0.34.0",
    "mlflow>=2.8.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    "catboost>=1.2.0",
    "streamlit>=1.35.0",
    "plotly>=5.22.0",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "httpx>=0.27.0",
]
data = [
    "dvc>=3.67.1",
    "dvc-gdrive>=3.0.1",
    "dvc-s3>=3.3.0",
]

[project.scripts]
run-pipeline = "rossmann_mlops.train_model:main"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
